requirements.txt
apparmor==3.1.7
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brlapi==0.8.5
btrfsutil==6.8
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-init==24.1
configobj==5.0.8
cryptography==42.0.5
dbus-python==1.3.2
distro==1.9.0
docopt==0.6.2
docutils==0.20.1
evdev==1.7.0
Flask==3.0.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
LibAppArmor==3.1.7
louis==3.29.0
lxml==5.1.0
MarkupSafe==2.1.5
moddb==0.10.0
netifaces==0.11.0
nftables==0.1
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.1
pillow==10.2.0
ply==3.11
psutil==5.9.8
pulsemixer==1.5.1
pycairo==1.26.0
pycparser==2.22
PyGObject==3.48.1
pyrate-limiter==3.6.0
pyrsistent==0.19.3
pyserial==3.5
python-dateutil==2.8.2
pytz==2024.1
pywal==3.3.0
PyYAML==6.0.1
ranger-fm==1.9.3
referencing==0.33.0
Reflector==2023.6.28.0.36.1
requests==2.31.0
rpds-py==0.17.1
six==1.16.0
soupsieve==2.5
toolz==0.12.1
trash-cli==0.23.11.10
typing_extensions==4.10.0
tzdata==2024.1
ueberzug==18.2.1
urllib3==1.26.18
validate==5.0.8
Werkzeug==3.0.1
